Week 3: the extension builds, loads, and makes a sound
Fifty objects drop in a real Godot 4.7.2 scene: one model loaded, ten voices
at the peak, nothing dropped, nothing stolen, -2.7 dBFS with the impacts
arriving between 0.8 and 1.4 seconds and ringing out after. Built and run
here, not assumed.

The demo found three things no unit test would have.

Nothing ever sounded, because ModalBody expected the user to call
read_contacts from _integrate_forces and the demo — written by the same
person who designed it — forgot. It reads the body's direct state itself now,
once a physics tick. The force-integration callback would be the other way
in, but RigidBody3D already owns that one and taking it would break the
body's own state sync.

Fifty marbles loaded fifty copies of one model. Cached by path.

And a rolling contact started a new voice every tick: 208 steals in five
seconds. The hint that identifies a continuing contact was the pool's own
voice id, which the physics thread can never learn, because the queue only
goes one way. The producer names the contact instead and the pool adopts the
name. Ten voices where there were sixty-four.

Also a plan correction: AudioStreamPlayback::_mix takes a float rate scale,
not the double section 6.2 has. With a double it silently fails to override —
it compiles as a new overload, the base class keeps returning silence, and
nothing says so.

The callback is now held to not allocating by a test rather than by reading
it: every global allocation is counted, and the count may not move while two
hundred blocks are mixed with voices starting throughout. A test that the
counter itself works sits beside it, or the first one would pass by being
broken.

Master gain is the one constant nothing physical decides — impulse arrives in
newton seconds and output leaves in full scale. Defaulted so a small object
dropped a metre or two lands near -12 dBFS, and exposed.

+++ godot/src/modal_stream.cpp
@@ -11,7 +11,7 @@ void ModalAudioStreamPlayback::_start(double) { playing_ = true; }
 void ModalAudioStreamPlayback::_stop() { playing_ = false; }
 bool ModalAudioStreamPlayback::_is_playing() const { return playing_; }
 
-int32_t ModalAudioStreamPlayback::_mix(AudioFrame* buffer, double, int32_t frames) {
+int32_t ModalAudioStreamPlayback::_mix(AudioFrame* buffer, float, int32_t frames) {
     // Real-time thread. No allocation, no locks, no logging, no Godot object
     // creation, and nothing that can throw.
     //
